Move prompt logprobs mode e2e test to its own module
Even at gpu_memory_utilization=0.05 the test cannot start next to the
module-scoped llm fixture: build 76618 shows only 0.72 GiB free on the
16 GiB CI GPU. Run it in a separate module instead, which pytest
collects before the fixture module, so the engines never coexist.
